Fix `l.rb` working in sub-directories of the git repo
This fixes `l.rb` working in sub-directories of the git repo by
adding the `--relative` flag to the `git diff` command. For
example, if you have the following folder structure:

    gitrepo
     - .git
     - subdirectory
        - .l.yml
        - files

and you are in the `subdirectory` folder, if you ran `l -c` to
detect changes in your `files` folder, it previously wouldn't see
any. This is because the `git diff` command returned results like
`gitrepo/subdirectory/files/something` and then the linter running
logic wouldn't think it's a valid file for it to run. The
`--relative` option makes it return `subdirectory/files/something`
which works as expected with the rest of the linter running logic.