(hotfix) allow for backwards-compatible linter configs
This allows the user to configure any linter attributes they like
in the `.l.yml` configs and not blow up building Linter instances
from the configs.

I noticed after I added the new `autocorrect_cmd` config. I added
it to a project without installing the latest version that
supported it and it caused the older version to raise an exception
about an unknown keyword arg. This fixes that issue and also the
generic issuer where a user enters a bad/unknown config attribute
on a linter.